+++ R/mrs_data_display.R
@@ -410,9 +410,9 @@ image.mrs_data <- function(x, xlim = NULL, mode = "re", col = NULL,
     graphics::axis(2, at = y_ticks, labels = F, col = NA, col.ticks = "red")
   }
   
-  if (!is.null(vline)) graphics::abline(v = vline, col = "white")
+  if (!is.null(vline)) graphics::abline(v = vline, col = "white", lty = 2)
   
-  if (!is.null(hline)) graphics::abline(h = hline, col = "white")
+  if (!is.null(hline)) graphics::abline(h = hline, col = "white", lty = 2)
   
   if (restore_def_par) graphics::par(.pardefault)
 }
@@ -665,7 +665,7 @@ stackplot.mrs_data <- function(x, xlim = NULL, mode = "re", x_units = NULL,
                                                    graphics::grid(nx = grid_nx,
                                                    ny = grid_ny)}, ...)
   
-  graphics::axis(1, lwd = 0, lwd.ticks = 1, at = pretty(xlim_labs))
+  graphics::axis(1, lwd = 0, lwd.ticks = 1, at = pretty(xlim_labs, 6))
   
   if (bty == "n") graphics::lines(xlim_labs, c(graphics::par("usr")[3],
                                                graphics::par("usr")[3]))